Re-add storage capability
Storage was not working after editing person and introducing
TelegramHandle and ModTutGroup classes.
Ordering of arguments for email and telegramHandle are not consistent.

Added JsonAdaptedModTutGroup class which is an edited copy of the
pre-existing JsonAdpatedTag class. Removed Tag related functions from
storage and replaced them with ModTutGroup instead. Reintroduced some
code which had been commented out.
Changed ordering of arguments for email and telegramHandle to be
consistent.