Change find command to allow searching by telehandle and email
Before, find command only allows for one predicate, which checks the
name.

Let's change it to take a list of predicates, and add some classes
similar to NameContainsKeywordsPredicates.

This allows the user to specify whether they want to search by name,
telehandle, email, or some combination of the three, adding more fields
the more specific they want the search to be.
This also allows for potential future expansion of find to include other
fields.